Bug fixes, manual adding items, fetch user content fixes, folder structure
