+++ src/server/controllers/menu/postAddManualPost.ts
@@ -1,8 +1,9 @@
 import { Request, Response } from 'express';
 import { context, settings, redis, reddit } from '@devvit/web/server';
+import { processPost } from '../../lib/processPost';
 
 interface PostMenuBody {
-  targetId: `t3_${string}`; 
+  targetId: `t3_${string}`;
   location: 'post'
 }
 
@@ -11,17 +12,42 @@ export const postAddManualPost = async (
   res: Response
 ): Promise<void> => {
   try {
-    let body: PostMenuBody = _req.body
-    let post = await reddit.getPostById(body.targetId)
+    const body: PostMenuBody = _req.body;
+    const post = await reddit.getPostById(body.targetId);
+
+    if (!post) {
+      res.status(404).json({
+        showToast: 'Post not found'
+      });
+      return;
+    }
+
+    // Check if post is from current subreddit
+    if (post.subredditName !== context.subredditName) {
+      res.status(400).json({
+        showToast: 'Cannot add posts from other subreddits'
+      });
+      return;
+    }
+
+    await processPost(post);
 
     res.json({
-      showToast: 'console logged body of post.)'
+      showToast: 'Post added successfully!'
     });
 
-  } catch (error) {
-    console.error(`[postFetchUserContent] Error:`, error);
-    res.status(400).json({
-      showToast: 'Processing failed. Please try again or check console logs.'
-    });
+  } catch (error: any) {
+    console.error(`[postAddManualPost] Error:`, error);
+
+    // Handle specific error cases
+    if (error?.message?.includes('already exists')) {
+      res.status(400).json({
+        showToast: 'Post already exists in the app'
+      });
+    } else {
+      res.status(400).json({
+        showToast: 'Failed to add post. Please try again.'
+      });
+    }
   }
 };
